InstituteApiController

+++ app/Http/Controllers/API/admin/InstituteApiController.php
@@ -133,7 +133,7 @@ class InstituteApiController extends Controller
                             $standardids .= 0;
                             $standardids = $standardidget->pluck('standard')->toArray();
                             $standard_array = Standard_model::whereIN('id', $standardids)
-                                ->get();
+                            ->get();
 
 
                             $standard = [];
@@ -380,35 +380,46 @@ class InstituteApiController extends Controller
                 $standard = $svaluee->standard;
                 $stream = $svaluee->stream;
 
+                $insfor = Institute_for_sub::where('institute_id',$lastInsertedId)
+                ->where('institute_for_id',$institute_for)->first();
+                if(empty($insfor)){
+                
+                    $createinstitutefor = Institute_for_sub::create([
+                        'user_id' => $request->input('user_id'),
+                        'institute_id' => $lastInsertedId,
+                        'institute_for_id' => $institute_for,
+                    ]);
 
-                $createinstitutefor = Institute_for_sub::create([
-                    'user_id' => $request->input('user_id'),
-                    'institute_id' => $lastInsertedId,
-                    'institute_for_id' => $institute_for,
-                ]);
-
-                if (!$createinstitutefor) {
-                    $instituteFordet = Institute_detail::where('id', $lastInsertedId)
-                        ->where('user_id', $request->input('user_id'))->first();
-                    $delt = $instituteFordet->delete();
+                    if (!$createinstitutefor) {
+                        $instituteFordet = Institute_detail::where('id', $lastInsertedId)
+                            ->where('user_id', $request->input('user_id'))->first();
+                        $delt = $instituteFordet->delete();
+                    }
                 }
 
-                $createboard = Institute_board_sub::create([
-                    'user_id' => $request->input('user_id'),
-                    'institute_id' => $lastInsertedId,
-                    'institute_for_id' => $institute_for,
-                    'board_id' => $board,
-                ]);
-
-                if (!$createboard) {
-                    $instituteFordet = Institute_detail::where('id', $lastInsertedId)
-                        ->where('user_id', $request->input('user_id'))->first();
-                    $instituteFordet->delete();
-
-                    $instituteForSub = Institute_for_sub::where('institute_id', $lastInsertedId)
-                        ->where('user_id', $request->input('user_id'))->delete();
+                $bordsubr = Institute_board_sub::where('institute_id',$lastInsertedId)
+                ->where('institute_for_id',$institute_for)
+                ->where('board_id',$board)->first();
+                if(empty($bordsubr)){
+                    $createboard = Institute_board_sub::create([
+                        'user_id' => $request->input('user_id'),
+                        'institute_id' => $lastInsertedId,
+                        'institute_for_id' => $institute_for,
+                        'board_id' => $board,
+                    ]);
+    
+                    if (!$createboard) {
+                        $instituteFordet = Institute_detail::where('id', $lastInsertedId)
+                            ->where('user_id', $request->input('user_id'))->first();
+                        $instituteFordet->delete();
+    
+                        $instituteForSub = Institute_for_sub::where('institute_id', $lastInsertedId)
+                            ->where('user_id', $request->input('user_id'))->delete();
+                    }
+    
                 }
 
+                
                 $createmedium = Medium_sub::create([
                     'user_id' => $request->input('user_id'),
                     'institute_id' => $lastInsertedId,
@@ -1729,12 +1740,14 @@ class InstituteApiController extends Controller
 
                     $boards[] = array(
                         'id' => $boardsdt->id,
-                        'name' => $boardsdt->name, 'medium' => $mediums
+                        'name' => $boardsdt->name, 
+                        'medium' => $mediums
                     );
                 }
                 $institute_fors[] = array(
                     'id' => $inst_forsd->id,
-                    'name' => $inst_forsd->name, 'boards' => $boards
+                    'name' => $inst_forsd->name,
+                     'boards' => $boards
                 );
             }
 
@@ -1770,7 +1783,7 @@ class InstituteApiController extends Controller
             'institute_id' => 'required',
             'user_id' => 'required',
             'exam_id' => 'required',
-            'batch_id' => 'required',
+            'batch_id'=> 'required',
         ]);
 
         if ($validator->fails()) {
@@ -1795,7 +1808,7 @@ class InstituteApiController extends Controller
             $exam_id = $request->exam_id;
             $batch_id = $request->batch_id;
             $examdt = Exam_Model::where('id', $exam_id)->first();
-
+            
             if (!empty($examdt)) {
 
                 $studentDT = Student_detail::join('users', 'users.id', '=', 'students_details.student_id')
@@ -1813,16 +1826,16 @@ class InstituteApiController extends Controller
                     })
                     ->whereRaw("FIND_IN_SET($examdt->subject_id, students_details.subject_id)")
                     ->select('students_details.*', 'users.firstname', 'users.lastname', 'standard.name as standardname')->get();
-
-                $studentsDET = [];
+                    
+                    $studentsDET = [];
                 foreach ($studentDT as $stddt) {
                     $subjectqy = Subject_model::where('id', $examdt->subject_id)->first();
                     $marksofstd = Marks_model::where('student_id', $stddt->student_id)->where('exam_id', $request->exam_id)->first();
                     $studentsDET[] = array(
                         'student_id' => $stddt->student_id,
                         'exam_id' => $request->exam_id,
-                        'batch_id' => $request->batch_id,
-                        'marks' => !empty($marksofstd->mark) ? (float)$marksofstd->mark : 0,
+                        'batch_id'=>$request->batch_id,
+                        'marks' => !empty($marksofstd->mark) ? (float)$marksofstd->mark : '',
                         'firstname' => $stddt->firstname,
                         'lastname' => $stddt->lastname,
                         'total_mark' => $examdt->total_mark,
